@@ -18,16 +18,16 @@
   "dependencies": {
     "@agm/core": "^3.0.0-beta.0",
     "@agm/snazzy-info-window": "^3.0.0-beta.0",
-    "@angular/animations": "^14.3.0",
-    "@angular/cdk": "^13.3.9",
-    "@angular/common": "^14.3.0",
-    "@angular/compiler": "^14.3.0",
-    "@angular/core": "^13.4.0",
-    "@angular/forms": "^14.3.0",
-    "@angular/material": "^13.3.9",
-    "@angular/platform-browser": "^14.3.0",
-    "@angular/platform-browser-dynamic": "^14.3.0",
-    "@angular/router": "^14.3.0",
+    "@angular/animations": "^15.2.9",
+    "@angular/cdk": "^14.2.7",
+    "@angular/common": "^15.2.9",
+    "@angular/compiler": "^15.2.9",
+    "@angular/core": "^15.2.9",
+    "@angular/forms": "^15.2.9",
+    "@angular/material": "^14.2.7",
+    "@angular/platform-browser": "^15.2.9",
+    "@angular/platform-browser-dynamic": "^15.2.9",
+    "@angular/router": "^15.2.9",
     "@apollo/client": "^3.3.20",
     "@cds/angular": "^5.4.1",
     "@cds/core": "^5.4.1",
@@ -69,15 +69,15 @@
   },
   "devDependencies": {
     "@angular-builders/custom-webpack": "^12.1.0",
-    "@angular-devkit/build-angular": "^13.3.11",
+    "@angular-devkit/build-angular": "^14.2.12",
     "@angular-eslint/builder": "^16.1.1",
     "@angular-eslint/eslint-plugin": "^16.1.1",
     "@angular-eslint/eslint-plugin-template": "^16.1.1",
     "@angular-eslint/schematics": "^16.1.1",
     "@angular-eslint/template-parser": "^16.1.1",
-    "@angular/cli": "^13.3.11",
-    "@angular/compiler-cli": "^14.3.0",
-    "@angular/language-service": "^14.3.0",
+    "@angular/cli": "^14.2.12",
+    "@angular/compiler-cli": "^15.2.9",
+    "@angular/language-service": "^15.2.9",
     "@graphql-codegen/cli": "^2.0.1",
     "@graphql-codegen/fragment-matcher": "2.0.0",
     "@graphql-codegen/typescript": "^1.22.4",
@@ -104,6 +104,6 @@
     "npm-run-all": "^4.1.5",
     "protractor": "~7.0.0",
     "ts-node": "~7.0.0",
-    "typescript": "^4.6.4"
+    "typescript": "^4.9.5"
   }
 }
